+++ examples/dotnet/SFKB_clientTests/SFKB_clientTest.cs
@@ -1,12 +1,8 @@
 using System;
 using System.IO;
 using System.Linq;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Text;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Newtonsoft.Json.Linq;
 using SFKB_API;
 
 namespace SFKB_clientTests
@@ -14,7 +10,6 @@ namespace SFKB_clientTests
     [TestClass]
     public class SfkbClientTest
     {
-        private const string ConfigLocation = @"../../../config.json";
         private static Client Client;
         private Guid DatasetId;
         private static readonly Guid WrongDatasetId = new Guid("2fa85f64-5717-4562-b3fc-2c963f66afa6");
@@ -25,36 +20,10 @@ namespace SFKB_clientTests
         [TestInitialize]
         public async Task InitAsync()
         {
-            var basicIdentification = File.Exists(ConfigLocation)
-                ? GetIdentificationFromConfigFile()
-                : GetIdentificationFromEnvironmentVariables();
-
-            var byteArray = Encoding.ASCII.GetBytes(basicIdentification);
-
-            var httpClient = new HttpClient();
-
-            var basicValue = Convert.ToBase64String(byteArray);
-
-            httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Basic", basicValue);
-
-            Client = new Client(httpClient);
+            Client = General.GetClientWithBasicAuthentication();
 
             await GetDatasetsAsync();
-        }
-
-        private static string GetIdentificationFromEnvironmentVariables()
-        {
-            return $"{Environment.GetEnvironmentVariable("api_user")}:{Environment.GetEnvironmentVariable("api_pass")}";
-        }
-
-        private static string GetIdentificationFromConfigFile()
-        {
-            string basicIdentification;
-            var config = JObject.Parse(File.ReadAllText(ConfigLocation));
-
-            basicIdentification = $"{config["user"].ToString()}:{config["password"].ToString()}";
-            return basicIdentification;
-        }
+        }        
 
         [TestMethod]
         public async Task TestGetDatasetsAsync()
@@ -62,26 +31,13 @@ namespace SFKB_clientTests
             var dataset = await GetDataset();
 
             Assert.IsTrue(dataset.Id == DatasetId, $"No dataset found with id {DatasetId}");
-        }
-
-        private Task<Dataset> GetDataset()
-        {
-            return Client.GetDatasetMetadataAsync(DatasetId);
-        }
-
-        private async Task GetDatasetsAsync()
-        {
-            var datasets = await Client.GetDatasetsAsync();
-
-            Assert.IsTrue(datasets.Count > 0, "No datasets returned");
-
-            DatasetId = datasets.FirstOrDefault( d => d.Name == Ar5DatasetName).Id;
-        }
+        }        
 
         [TestMethod]
         public void TestGetDatasetWithWrongId()
         {
             Exception exception = null;
+
             try
             {
                 var result = Client.GetDatasetMetadataAsync(WrongDatasetId).Result;
@@ -110,15 +66,49 @@ namespace SFKB_clientTests
             await ReplaceByLokalIdAsync(DatasetId, Ar5GrenseFeatureLokalId);
         }
 
+        [TestMethod]
+        public async Task TestGetLockedFeaturesByLokalIdAsync()
+        {
+            var locking = General.GetLocking();
+
+            await LockAndSaveFeatureByLokalIdAsync(Ar5FlateFeatureLokalId, locking);
+
+            await DeleteLocks(DatasetId, locking);
+        }
+
+        [TestMethod]
+        public async Task TestGetFeaturesByLokalIdAsync()
+        {
+            await LockAndSaveFeatureByLokalIdAsync(Ar5GrenseFeatureLokalId, null);
+        }
+
+        private Task<Dataset> GetDataset()
+        {
+            return Client.GetDatasetMetadataAsync(DatasetId);
+        }
+
+        private async Task GetDatasetsAsync()
+        {
+            var datasets = await Client.GetDatasetsAsync();
+
+            Assert.IsTrue(datasets.Count > 0, "No datasets returned");
+
+            DatasetId = datasets.FirstOrDefault(d => d.Name == Ar5DatasetName).Id;
+        }
+
         private async Task ReplaceByLokalIdAsync(Guid datasetId, string lokalId)
         {
-            var locking = GetLocking();
+            var locking = General.GetLocking();
 
             var tempFile = await LockAndSaveFeatureByLokalIdAsync(lokalId, locking);
 
             var datasetLocks = await Client.GetDatasetLocksAsync(datasetId, locking);
 
-            var lockedLokalIds = datasetLocks.SelectMany(l => l.Features.Select(f => f.Lokalid)).ToList();
+            Assert.IsTrue(datasetLocks?.Count() > 0, $"No dataset found for datasetId {datasetId}");
+
+            var lockedLokalIds = datasetLocks.SelectMany(l => l?.Features?.Select(f => f.Lokalid))?.ToList();
+
+            Assert.IsTrue(lockedLokalIds!= null && lockedLokalIds.Count > 0, $"No features locked for datasetId {datasetId} and lokalId {lokalId}");
 
             var wfsReplaceFile = Wfs.CreateReplaceWrappingForFeatures(tempFile, lockedLokalIds);
 
@@ -130,42 +120,6 @@ namespace SFKB_clientTests
             };
 
             await DeleteLocks(datasetId, locking);
-        }
-
-        //[TestMethod]
-        //public void TestGetFeaturesWithBbox()
-        //{
-        //    var fileResponse = Client.GetDatasetFeaturesAsync(DatasetId, null, GetExampleBbox(), null).Result;
-
-        //    WriteStreamToDisk(fileResponse);
-        //}
-
-        private static string WriteStreamToDisk(FileResponse fileResponse)
-        {
-            var tempFile = Path.GetTempFileName();
-
-            using (var fileStream = File.OpenWrite(tempFile))
-            {
-                fileResponse.Stream.CopyTo(fileStream);
-            }
-
-            Assert.IsTrue(File.Exists(tempFile), "No features saved to disk");
-
-            var fileInfo = new FileInfo(tempFile);
-
-            Assert.IsTrue(fileInfo.Length > 0, "Empty file saved to disk");
-            
-            return tempFile;
-        }
-
-        [TestMethod]
-        public async Task TestGetLockedFeaturesByLokalIdAsync()
-        {
-            var locking = GetLocking();
-
-            await LockAndSaveFeatureByLokalIdAsync(Ar5FlateFeatureLokalId, locking);
-
-            await DeleteLocks(DatasetId, locking);
         }
 
         private async Task DeleteLocks(Guid datasetId, Locking locking)
@@ -181,13 +135,7 @@ namespace SFKB_clientTests
         {
             var fileResponse = await Client.GetDatasetFeaturesAsync(DatasetId, locking, null, GetLokalIdQuery(lokalId));
 
-            return WriteStreamToDisk(fileResponse);
-        }
-
-        [TestMethod]
-        public async Task TestGetFeaturesByLokalIdAsync()
-        {
-            await LockAndSaveFeatureByLokalIdAsync(Ar5GrenseFeatureLokalId, null);
+            return General.WriteStreamToDisk(fileResponse);
         }
 
         private string GetLokalIdQuery(string lokalid)
@@ -212,9 +160,12 @@ namespace SFKB_clientTests
         //    return new StreamReader(fileName).BaseStream;
         //}
 
-        private static Locking GetLocking()
-        {
-            return new Locking { Type = LockingType.User_lock };
-        }
+        //[TestMethod]
+        //public void TestGetFeaturesWithBbox()
+        //{
+        //    var fileResponse = Client.GetDatasetFeaturesAsync(DatasetId, null, GetExampleBbox(), null).Result;
+
+        //    WriteStreamToDisk(fileResponse);
+        //}
     }
 }
